Refactor Postgres tests into separate modules
